+++ CHANGES.md
@@ -3,7 +3,9 @@
 ## X.Y.Z (YYYY-MM-DD)
 
 ## 0.4. (2021-07-XX)
+- `stack` accepts **kwargs in order to pass options to `rioxarray.to_raster()`
 - [CI] BUG: Fixing network directories with pathlib
+- Fixing some type hints
 
 ## 0.4.2 (2021-07-01)